requirements.txt
beautifulsoup4==4.8.1
cachetools==4.0.0
Click==7.0
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.1
gcloud==0.17.0
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.6.0
httplib2==0.15.0
idna==2.8
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.10.3
jws==0.1.3
lxml==4.4.2
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.4.5
oauth2client==3.0.0
Pillow==6.2.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
PyYAML==5.2
requests==2.11.1
requests-file==1.4.3
requests-toolbelt==0.7.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
tinysegmenter==0.3
tldextract==2.2.2
uritemplate==3.0.1
urllib3==1.25.7
Werkzeug==0.16.0
